interp,expand/test: fix GOARCH=386 build + windows glob test path separators
- os_stat_linux.go: Stat_t timespec fields are int32 on 32-bit (GOARCH=386)
  but int64 on amd64; convert explicitly for time.Unix (fixes the 386 build).
- expand_test: skip TestFieldsBackslashEscapedGlobPath +
  TestGlobLiteralUnreadableIntermediateDir on windows, where backslash is the
  path separator (not a glob escape) and glob results use \ — the unix-path
  assertions don't apply.

Both surfaced once CI's linux/windows jobs ran again (masked by the earlier
syscall.Getsid compile break).

+++ interp/os_stat_linux.go
@@ -11,9 +11,11 @@ import (
 )
 
 func unixStatMtime(s *syscall.Stat_t) time.Time {
-	return time.Unix(s.Mtim.Sec, s.Mtim.Nsec)
+	// Stat_t timespec fields are int64 on 64-bit but int32 on 32-bit
+	// (e.g. GOARCH=386), so convert explicitly for time.Unix (wants int64).
+	return time.Unix(int64(s.Mtim.Sec), int64(s.Mtim.Nsec))
 }
 
 func unixStatAtime(s *syscall.Stat_t) time.Time {
-	return time.Unix(s.Atim.Sec, s.Atim.Nsec)
+	return time.Unix(int64(s.Atim.Sec), int64(s.Atim.Nsec))
 }
